+++ baybe/searchspace/continuous.py
@@ -2,7 +2,9 @@
 
 from __future__ import annotations
 
+import warnings
 from collections.abc import Collection, Sequence
+from itertools import chain
 from typing import TYPE_CHECKING, Any, cast
 
 import numpy as np
@@ -10,8 +12,13 @@ import pandas as pd
 from attr import define, field
 
 from baybe.constraints import (
+    ContinuousCardinalityConstraint,
     ContinuousLinearEqualityConstraint,
     ContinuousLinearInequalityConstraint,
+)
+from baybe.constraints.base import ContinuousNonlinearConstraint
+from baybe.constraints.validation import (
+    validate_cardinality_constraints_are_nonoverlapping,
 )
 from baybe.parameters import NumericalContinuousParameter
 from baybe.parameters.base import ContinuousParameter
@@ -24,6 +31,8 @@ from baybe.utils.numerical import DTypeFloatNumpy
 
 if TYPE_CHECKING:
     from baybe.searchspace.core import SearchSpace
+
+_MAX_CARDINALITY_SAMPLING_ATTEMPTS = 10_000
 
 
 @define
@@ -38,17 +47,22 @@ class SubspaceContinuous(SerialMixin):
     parameters: tuple[NumericalContinuousParameter, ...] = field(
         converter=to_tuple, validator=lambda _, __, x: validate_parameter_names(x)
     )
-    """The list of parameters of the subspace."""
+    """The parameters of the subspace."""
 
     constraints_lin_eq: tuple[ContinuousLinearEqualityConstraint, ...] = field(
         converter=to_tuple, factory=tuple
     )
-    """List of linear equality constraints."""
+    """Linear equality constraints."""
 
     constraints_lin_ineq: tuple[ContinuousLinearInequalityConstraint, ...] = field(
         converter=to_tuple, factory=tuple
     )
-    """List of linear inequality constraints."""
+    """Linear inequality constraints."""
+
+    constraints_nonlin: tuple[ContinuousNonlinearConstraint, ...] = field(
+        converter=to_tuple, factory=tuple
+    )
+    """Nonlinear constraints."""
 
     def __str__(self) -> str:
         if self.is_empty:
@@ -63,9 +77,13 @@ class SubspaceContinuous(SerialMixin):
         ineq_constraints_list = [
             constr.summary() for constr in self.constraints_lin_ineq
         ]
+        nonlin_constraints_list = [
+            constr.summary() for constr in self.constraints_nonlin
+        ]
         param_df = pd.DataFrame(param_list)
         lin_eq_constr_df = pd.DataFrame(eq_constraints_list)
         lin_ineq_constr_df = pd.DataFrame(ineq_constraints_list)
+        cardinality_constr_df = pd.DataFrame(nonlin_constraints_list)
 
         # Put all attributes of the continuous class in one string
         continuous_str = f"""{start_bold}Continuous Search Space{end_bold}
@@ -73,9 +91,28 @@ class SubspaceContinuous(SerialMixin):
             \n{start_bold}List of Linear Equality Constraints{end_bold}
             \r{pretty_print_df(lin_eq_constr_df)}
             \n{start_bold}List of Linear Inequality Constraints{end_bold}
-            \r{pretty_print_df(lin_ineq_constr_df)}"""
+            \r{pretty_print_df(lin_ineq_constr_df)}
+            \n{start_bold}List of Cardinality Constraints{end_bold}
+            \r{pretty_print_df(cardinality_constr_df)}"""
 
         return continuous_str.replace("\n", "\n ").replace("\r", "\r ")
+
+    @property
+    def constraints_cardinality(self) -> tuple[ContinuousCardinalityConstraint, ...]:
+        """Cardinality constraints."""
+        return tuple(
+            c
+            for c in self.constraints_nonlin
+            if isinstance(c, ContinuousCardinalityConstraint)
+        )
+
+    @constraints_nonlin.validator
+    def _validate_constraints_nonlin(self, _, __) -> None:
+        """Validate nonlinear constraints."""
+        # Note: The passed constraints are accessed indirectly through the property
+        validate_cardinality_constraints_are_nonoverlapping(
+            self.constraints_cardinality
+        )
 
     def to_searchspace(self) -> SearchSpace:
         """Turn the subspace into a search space with no discrete part."""
@@ -177,6 +214,25 @@ class SubspaceContinuous(SerialMixin):
             return np.empty((2, 0), dtype=DTypeFloatNumpy)
         return np.stack([p.bounds.to_ndarray() for p in self.parameters]).T
 
+    def _drop_parameters(self, parameter_names: Collection[str]) -> SubspaceContinuous:
+        """Create a copy of the subspace with certain parameters removed.
+
+        Args:
+            parameter_names: The names of the parameter to be removed.
+
+        Returns:
+            The reduced subspace.
+        """
+        return SubspaceContinuous(
+            parameters=[p for p in self.parameters if p.name not in parameter_names],
+            constraints_lin_eq=[
+                c._drop_parameters(parameter_names) for c in self.constraints_lin_eq
+            ],
+            constraints_lin_ineq=[
+                c._drop_parameters(parameter_names) for c in self.constraints_lin_ineq
+            ],
+        )
+
     def transform(
         self,
         data: pd.DataFrame,
@@ -195,28 +251,70 @@ class SubspaceContinuous(SerialMixin):
         return comp_rep
 
     def samples_random(self, n_points: int = 1) -> pd.DataFrame:
-        """Get random point samples from the continuous space.
+        """Deprecated!"""  # noqa: D401
+        warnings.warn(
+            f"The method '{SubspaceContinuous.samples_random.__name__}' "
+            f"has been deprecated and will be removed in a future version. "
+            f"Use '{SubspaceContinuous.sample_uniform.__name__}' instead.",
+            DeprecationWarning,
+        )
+        return self.sample_uniform(n_points)
+
+    def sample_uniform(self, batch_size: int = 1) -> pd.DataFrame:
+        """Draw uniform random parameter configurations from the continuous space.
 
         Args:
-            n_points: Number of points that should be sampled.
+            batch_size: The number of parameter configurations to be sampled.
 
         Returns:
-            A dataframe containing the points as rows with columns corresponding to the
-            parameter names.
+            A dataframe containing the parameter configurations as rows with columns
+            corresponding to the parameter names.
+
+        Raises:
+            ValueError: If the subspace contains unsupported nonlinear constraints.
         """
+        if not all(
+            isinstance(c, ContinuousCardinalityConstraint)
+            for c in self.constraints_nonlin
+        ):
+            raise ValueError(
+                f"Currently, only nonlinear constraints of type "
+                f"'{ContinuousCardinalityConstraint.__name__}' are supported."
+            )
+
         if not self.parameters:
-            return pd.DataFrame()
+            return pd.DataFrame(index=pd.RangeIndex(0, batch_size))
+
+        if (
+            len(self.constraints_lin_eq) == 0
+            and len(self.constraints_lin_ineq) == 0
+            and len(self.constraints_cardinality) == 0
+        ):
+            return self._sample_from_bounds(batch_size, self.param_bounds_comp)
+
+        if len(self.constraints_cardinality) == 0:
+            return self._sample_from_polytope(batch_size, self.param_bounds_comp)
+
+        return self._sample_from_polytope_with_cardinality_constraints(batch_size)
+
+    def _sample_from_bounds(self, batch_size: int, bounds: np.ndarray) -> pd.DataFrame:
+        """Draw uniform random samples over a hyperrectangle-shaped space."""
+        points = np.random.uniform(
+            low=bounds[0, :], high=bounds[1, :], size=(batch_size, len(self.parameters))
+        )
+
+        return pd.DataFrame(points, columns=self.param_names)
+
+    def _sample_from_polytope(
+        self, batch_size: int, bounds: np.ndarray
+    ) -> pd.DataFrame:
+        """Draw uniform random samples from a polytope."""
         import torch
         from botorch.utils.sampling import get_polytope_samples
 
-        # TODO Revisit: torch and botorch here are actually only necessary if there
-        # are constraints. If there are none and the lists are empty we can just sample
-        # without the get_polytope_samples, which means torch and botorch
-        # wouldn't be needed.
-
         points = get_polytope_samples(
-            n=n_points,
-            bounds=torch.from_numpy(self.param_bounds_comp),
+            n=batch_size,
+            bounds=torch.from_numpy(bounds),
             equality_constraints=[
                 c.to_botorch(self.parameters) for c in self.constraints_lin_eq
             ],
@@ -224,32 +322,98 @@ class SubspaceContinuous(SerialMixin):
                 c.to_botorch(self.parameters) for c in self.constraints_lin_ineq
             ],
         )
-
         return pd.DataFrame(points, columns=self.param_names)
 
+    def _sample_from_polytope_with_cardinality_constraints(
+        self, batch_size: int
+    ) -> pd.DataFrame:
+        """Draw random samples from a polytope with cardinality constraints."""
+        if not self.constraints_cardinality:
+            raise RuntimeError(
+                f"This method should not be called without any constraints of type "
+                f"'{ContinuousCardinalityConstraint.__name__}' in place. "
+                f"Use '{SubspaceContinuous._sample_from_bounds.__name__}' "
+                f"or '{SubspaceContinuous._sample_from_polytope.__name__}' instead."
+            )
+
+        # List to store the created samples
+        samples: list[pd.DataFrame] = []
+
+        # Counter for failed sampling attempts
+        n_fails = 0
+
+        while len(samples) < batch_size:
+            # Randomly set some parameters inactive
+            inactive_params_sample = self._sample_inactive_parameters(1)[0]
+
+            # Remove the inactive parameters from the search space
+            subspace_without_cardinality_constraint = self._drop_parameters(
+                inactive_params_sample
+            )
+
+            # Sample from the reduced space
+            try:
+                sample = subspace_without_cardinality_constraint.sample_uniform(1)
+                samples.append(sample)
+            except ValueError:
+                n_fails += 1
+
+            # Avoid infinite loop
+            if n_fails >= _MAX_CARDINALITY_SAMPLING_ATTEMPTS:
+                raise RuntimeError(
+                    f"The number of failed sampling attempts has exceeded the limit "
+                    f"of {_MAX_CARDINALITY_SAMPLING_ATTEMPTS}. "
+                    f"It appears that the feasible region of the search space is very "
+                    f"small. Please review the search space constraints."
+                )
+
+        # Combine the samples and fill in inactive parameters
+        parameter_names = [p.name for p in self.parameters]
+        return (
+            pd.concat(samples, ignore_index=True)
+            .reindex(columns=parameter_names)
+            .fillna(0.0)
+        )
+
+    def _sample_inactive_parameters(self, batch_size: int = 1) -> list[set[str]]:
+        """Sample inactive parameters according to the given cardinality constraints."""
+        inactives_per_constraint = [
+            con.sample_inactive_parameters(batch_size)
+            for con in self.constraints_cardinality
+        ]
+        return [set(chain(*x)) for x in zip(*inactives_per_constraint)]
+
     def samples_full_factorial(self, n_points: int = 1) -> pd.DataFrame:
-        """Get random point samples from the full factorial of the continuous space.
+        """Deprecated!"""  # noqa: D401
+        warnings.warn(
+            f"The method '{SubspaceContinuous.samples_full_factorial.__name__}' "
+            f"has been deprecated and will be removed in a future version. "
+            f"Use '{SubspaceContinuous.sample_from_full_factorial.__name__}' instead.",
+            DeprecationWarning,
+        )
+        return self.sample_from_full_factorial(n_points)
+
+    def sample_from_full_factorial(self, batch_size: int = 1) -> pd.DataFrame:
+        """Draw parameter configurations from the full factorial of the space.
 
         Args:
-            n_points: Number of points that should be sampled.
+            batch_size: The number of parameter configurations to be sampled.
 
         Returns:
-            A dataframe containing the points as rows with columns corresponding to the
-            parameter names.
+            A dataframe containing the parameter configurations as rows with columns
+            corresponding to the parameter names.
 
         Raises:
             ValueError: If there are not enough points to sample from.
         """
-        full_factorial = self.full_factorial
-
-        if len(full_factorial) < n_points:
+        if len(full_factorial := self.full_factorial) < batch_size:
             raise ValueError(
-                f"You are trying to sample {n_points} points from the full factorial of"
-                f" the continuous space bounds, but it has only {len(full_factorial)} "
-                f"points."
+                f"You are trying to sample {batch_size} points from the full factorial "
+                f"of the continuous space bounds, but it has only "
+                f"{len(full_factorial)} points."
             )
 
-        return full_factorial.sample(n=n_points).reset_index(drop=True)
+        return full_factorial.sample(n=batch_size).reset_index(drop=True)
 
     @property
     def full_factorial(self) -> pd.DataFrame: