+++ baybe/searchspace/core.py
@@ -15,11 +15,8 @@ from baybe.constraints import (
     ContinuousLinearInequalityConstraint,
     validate_constraints,
 )
-from baybe.constraints.base import Constraint
-from baybe.parameters import (
-    SubstanceEncoding,
-    TaskParameter,
-)
+from baybe.constraints.base import Constraint, ContinuousNonlinearConstraint
+from baybe.parameters import SubstanceEncoding, TaskParameter
 from baybe.parameters.base import Parameter
 from baybe.searchspace.continuous import SubspaceContinuous
 from baybe.searchspace.discrete import (
@@ -153,6 +150,9 @@ class SearchSpace(SerialMixin):
                 for c in constraints
                 if isinstance(c, ContinuousLinearInequalityConstraint)
             ],
+            constraints_nonlin=[
+                c for c in constraints if isinstance(c, ContinuousNonlinearConstraint)
+            ],
         )
 
         return SearchSpace(discrete=discrete, continuous=continuous)
@@ -214,6 +214,7 @@ class SearchSpace(SerialMixin):
             *self.discrete.constraints,
             *self.continuous.constraints_lin_eq,
             *self.continuous.constraints_lin_ineq,
+            *self.continuous.constraints_nonlin,
         )
 
     @property